re-implemented GurobiBackend to use only C API
On MACs, the C++ Gurobi interface causes too much trouble: The shared
library was compiled against libstdc++, but the default on MACs is
libc++. Consequently, a lot of other libraries (e.g., boost) have to be
recompiled to be binary compatible with the Gurobi C++ library.

By switching to the C API, we don't need the Gurobi C++ library anymore
and got rid of this problem.